requirements.txt
scikit-learn==0.20.2
scipy==1.2.0
sklearn-crfsuite==0.3.6
tensorflow==1.13.2
word2number==1.1
rasa.nlu==0.14.3
tensorflow-hub==0.2.0
spacy==2.0.18
editdistance==0.5.2
